[anlytc] + study param. inp-fcn, exmple triangular
Added functions for calculating results of an input signal which varies its
shape depending on a parameter. With this, the shape of the leaf depending on
crest and form factor can be studied.

FIXME: For dirac-like impulses, the time integration struggles. For this,
incorporate a quality check of the solution which increases step size if
neccessary.

On branch feature/anlytc
Changes to be committed:
      new file:   sample/plot_study_results.m
      new file:   sample/study_signal_group.m
      new file:   sample/study_tri.m