[anlytc] + theoretical peak calc (C,F) |--> (x,y)
Added function that calculates the peak (x,y) of a l_shape function with form
and crest factor:
    calc_peak:      general calculation
    add_theo_peak:  aplication to data
    main:           adding theo peak in main fcn

Description update and typo correction in other files.

Result of analytical calculation: C and F are not sufficient for a description
of the leaf form, although it gives a clear hint. Calculation only valid for
l_shape function, not for others. New idea: Sorted load profile.

On branch feature/anlytc
Changes to be committed:
      new file:   project/add_theo_peak.m
      new file:   project/auxiliary/calc_peak.m
      modified:   project/main.m
      modified:   project/plot_transformed.m
      modified:   project/transform_and_peak.m
      modified:   sample/study_distorted_sin.m