[interpt] + interpoint calculation, * main, hybrid
Added functionality to calculate a pair of base/peak storages within the leaf.
For this, a ratio inter was defined, determining which amount of signal is
handed to the peak storage before it goes to the base storage. This leads to an
increase of energy of the peak storage, while retaining the original power cut
ratio. With this, every pair of points within the leaf can be reached but it
can not be predicted which pair will be met with a certain inter value.

calc_hybrid_storage: * base and signal fcn def, + fcn input args (inter added)
main: + fcn input args (inter added), call to calc_hybrid_storage complemented

On branch feature/interpt
Changes to be committed:
      modified:   project/calc_hybrid_storage.m
      modified:   project/main.m