[dev] + results of tests w/ peak/base swapped
A few result .fig which analyze the behaviour of different input signals.
It studies the differences between mirrored signals and point mirrored signals
(for some signals) and shows that the peak storage can be larger than the base
storage, if input signal temporarily changes sign.

On branch develop
Changes to be committed:
      new file:   sample/signal_v1_mirrored.fig
      new file:   sample/signal_v1_pointsym.fig
      new file:   sample/signal_v2_mirrored.fig
      new file:   sample/signal_v2_pointsym.fig