[reload] * ode/dae strcture, 1st half period works
Converted dae to ode which leads to an operational strategy which works as
expected and does not show some unexpected behaviour when the the input is
separated to the storages. This _may_ be caused by some unknown handling of
algebraic constraints.

Operational strategy is only implemented correctly for the first half of the
period and needs to be extended.

On branch feature/reload
Changes to be committed:
      modified:   project/auxiliary/op_strat_reload_dim.m
      modified:   project/calc_hybrid_w_reload.m