[reload] * op strat for 2nd half period
Modified operational strategy in a way that the second half period is simulated
correctly (leaving aside the errors at the end of the period which are not
interesting at the moment). Operational strategy needs a priori knowledge of
dimensions (which is only valid for the investigated signal - double triangle
w/ amplitude of one and period of 2pi).
Restored ODE structure, instead of DAE structure (as mentioned in previous
commit).

On branch feature/reload
Changes to be committed:
      modified:   project/auxiliary/op_strat_reload_dim.m
      modified:   project/calc_hybrid_w_reload.m