[refctr] * enhanced plot sim, * mv fhandle calc
Enhanced the plotting capabilites of plot_simulation. Moved fhandle calculation
of sim_operation to separate function, enhanced output of sim_operation.

Please enter the commit message for your changes. Lines starting
with '#' will be ignored, and an empty message aborts the commit.
On branch feature/refctr
Changes to be committed:
      modified:   src/_sim_operation/control_factory.m
      new file:   src/_sim_operation/sim_operation_fhandle.m
      modified:   src/hybridset.m
      modified:   src/plot_operation.m
      modified:   src/sim_operation.m