dev+ added hybrid_simple_call and test for this
Added a function hybrid_simple_call that automates some steps for single
discharge step signals. It prepares the signal and calls the hybrid
method with reasonable hybridset values and saves the data in a specific
format that makes it easy to import/export into python.

On branch develop
Changes to be committed:
      modified:   hybrid.m
      new file:   sample/hybrid_simple_call.m
      new file:   sample/test_simple_call.m